requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
django-crispy-forms==1.14.0
django-phone-field==1.8.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
